Move some functions to util.go
