+++ util.go
@@ -7,6 +7,7 @@ import (
 	"io/ioutil"
 	"os"
 	"path/filepath"
+	"github.com/golang/go/src/regexp"
 )
 
 func zipDirectory(directory string) (*bytes.Buffer, error) {
@@ -63,4 +64,73 @@ func find(targetDir string) ([]string, error) {
 	}
 
 	return paths, nil
+}
+
+func cleanTempDirectory(directory string) error {
+	if err := os.RemoveAll(directory); err != nil {
+		return err
+	}
+	return nil
+}
+
+func extractInstallParameter(url string) (uri string, dir string, target_dir string, branch string) {
+	r := regexp.MustCompile(`^(https://([^/]+?)/([^/]+?)/([^/@]+?))(/([^@]+))?(@([^/]+))?$`)
+	group := r.FindAllStringSubmatch(url, -1)
+	uri = group[0][1]
+	dir = group[0][4]
+	target_dir = group[0][6]
+	branch = group[0][8]
+	if branch == "" {
+		branch = "master"
+	}
+	return
+}
+
+func loadInstallUrls(args cli.Args) ([]string, error) {
+	urls := []string{}
+	if c.Config.PackageFile != "" {
+		packageFile, err := readPackageFile(c.Config.PackageFile)
+		if err != nil {
+			return nil, err
+		}
+		for _, pkg := range packageFile.Packages {
+			url, err := convertToUrl(pkg)
+			if err != nil {
+				return nil, err
+			}
+			urls = append(urls, url)
+		}
+	} else {
+		url, err := convertToUrl(args.First())
+		if err != nil {
+			return nil, err
+		}
+		urls = []string{url}
+	}
+	return urls, nil
+}
+
+func convertToUrl(target string) (string, error) {
+	if target == "" {
+		return "", errors.New("Repository not specified")
+	}
+	url := target
+	r := regexp.MustCompile(`^[^/]+?/[^/@]+?(/[^@]+?)?(@[^/]+)?$`)
+	if r.MatchString(url) {
+		url = DEFAULT_REPOSITORY + "/" + url
+	}
+	return "https://" + url, nil
+}
+
+func readPackageFile(packageFileStr string) (*PackageFile, error) {
+	packageFile := PackageFile{}
+	readBody, err := ioutil.ReadFile(packageFileStr)
+	if err != nil {
+		return nil, err
+	}
+	err = yaml.Unmarshal([]byte(readBody), &packageFile)
+	if err != nil {
+		return nil, err
+	}
+	return &packageFile, nil
 }
